+++ examples/scrape_one.rs
@@ -5,7 +5,7 @@
 
 use google_maps_scraper::{MapsScraper, ScraperConfig};
 
-#[tokio::main]
+#[tokio::main(flavor = "multi_thread")]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     tracing_subscriber::fmt()
         .with_env_filter(